+++ zcash_client_backend/src/data_api/wallet/input_selection.rs
@@ -513,7 +513,7 @@ where
                 RecipientAddress::Transparent(addr) => {
                     push_transparent(*addr);
                 }
-                RecipientAddress::Shielded(_) => {
+                RecipientAddress::Sapling(_) => {
                     push_sapling();
                 }
                 RecipientAddress::Unified(addr) => {