+++ zcash_client_sqlite/src/chain.rs
@@ -520,7 +520,7 @@ mod tests {
 
         // We can spend the received notes
         let req = TransactionRequest::new(vec![Payment {
-            recipient_address: RecipientAddress::Shielded(dfvk.default_address().1),
+            recipient_address: RecipientAddress::Sapling(dfvk.default_address().1),
             amount: NonNegativeAmount::const_from_u64(110_000),
             memo: None,
             label: None,